+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+canary:
+  - now automatically infer the port from the protocol
+
 0.0.1:
   - basic engine (not yet working)